move tests, add success test
